+++ tests/testthat/test-mass-polr.R
@@ -7,6 +7,7 @@ skip_if_not_installed("MASS")
 library(MASS)
 
 fit <- polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing)
+fit2 <- polr(Sat ~ Freq, data = housing)
 
 test_that("MASS::polr tidier arguments", {
   check_arguments(tidy.polr)
@@ -18,12 +19,15 @@ test_that("tidy.polr", {
   
   td <- tidy(fit, quick = TRUE)
   td2 <- tidy(fit, conf.int = TRUE, exponentiate = TRUE)
+  td3 <- tidy(fit2, conf.int = TRUE, exponentiate = TRUE)
   
   check_tidy_output(td, strict = FALSE)
   check_tidy_output(td2, strict = FALSE)
+  check_tidy_output(td3, strict = FALSE)
   
   check_dims(td, expected_cols = 3)
   check_dims(td2, expected_cols = 7)
+  check_dims(td3, expected_cols = 7)
 })
 
 test_that("glance.polr", {